requirements.txt
boto3==1.16.26
botocore==1.19.26
jmespath==0.10.0
numpy==1.19.4
pandas==1.1.4
pyarrow==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
